--- pyAitu/commands/__init__.py
@@ -1,11 +1,0 @@
-from .command import Command
-from .reply_command import ReplyCommand
-from .quick_button_command import QuickButtonCommand
-from .inline_command import InlineCommand
-
-__all__ = [
-    Command,
-    ReplyCommand,
-    QuickButtonCommand,
-    InlineCommand
-]
