+++ pyAitu/models/command/uiState.py
@@ -1,4 +1,4 @@
-from pyAitu.utils.strings import SHOW_SHARE_CONTACT_BUTTON,\
+from ...utils.strings import SHOW_SHARE_CONTACT_BUTTON,\
     SHOW_RECORD_AUDIO_BUTTON, SHOW_GALLERY_BUTTON, SHOW_CAMERA_BUTTON, REPLY_KEYBOARD, QUICK_BUTTON_COMMANDS
