+++ pyAitu/models/media/__init__.py
@@ -1,0 +1,7 @@
+from .media import Media
+from .image import Image
+
+__all__ = [
+    Media,
+    Image
+]
