+++ pyAitu/models/media/image.py
