+++ pyAitu/models/peer/__init__.py
@@ -1,0 +1,9 @@
+from .recipient import Recipient
+from .peer import Peer
+from .peer_user import PeerUser
+
+__all__ = [
+    Recipient,
+    Peer,
+    PeerUser
+]
