+++ pyAitu/models/peer/peer_user.py
@@ -1,4 +1,4 @@
-from pyAitu.models.peer import Peer
+from .peer import Peer
 
 
 class PeerUser(Peer):
