+++ pyAitu/models/update/__init__.py
@@ -1,0 +1,12 @@
+from .update import Update
+from .quick_button_selected import QuickButtonSelected
+from .inline_command_selected import InlineCommandSelected
+from .message import Message, ContentType
+
+__all__ = [
+    Update,
+    QuickButtonSelected,
+    InlineCommandSelected,
+    Message,
+    ContentType
+]
